feat: guarda-corpo WIP + build multi-target Revit 2023/2026
- Adiciona feature de geração de guarda-corpo metálico (WIP):
  GenerateRailCommand, RailWindow, RailViewModel, RailCalculator,
  PostBuilder, HandrailBuilder, InfillBuilder (stub), SagaLog,
  EnumToBoolConverter, modelos RailConfig/RailDefinition/RailSegment/BarConfig

- Multi-target net48 (Revit 2023) + net8.0-windows (Revit 2026):
  um único build gera as duas DLLs e faz deploy nas pastas corretas

- Compat Revit 2025+: ElementId.IntegerValue removido, substituído por
  ElementIdExtensions.GetId() com #if NET8_0_OR_GREATER

- ConversionResult.ElementId: int → long (alinha com ElementId.Value)

- Melhorias na janela de escada (StairWindow, StairViewModel)
- .gitignore: exclui *.rvt, *.ifc e arquivos gerados pelo Revit

+++ SAGAStructuralTools/Core/Models/ConversionResult.cs
@@ -9,7 +9,7 @@ namespace SAGAStructuralTools.Core.Models
 
     public class ConversionResult
     {
-        public int              ElementId    { get; set; }
+        public long             ElementId    { get; set; }
         public string           OriginalName { get; set; }
         public ConversionStatus Status       { get; set; }
         public string           Message      { get; set; }
